requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
geoalchemy2==0.14.1
shapely==2.0.6
